Sharikly-main/backend/requirements.txt
asgiref==3.10.0
boto3==1.35.0
botocore==1.35.99
Django==5.2.7
django-anymail==11.0.0
django-cors-headers==4.9.0
django-storages==1.14.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.27.2
jmespath==1.0.1
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.6.2
requests==2.32.3
